# perf: Reuse queue allocation in `maximum_matching` main loop