A minor change that reduces the number of allocations in `maximum_matching` main loop and slightly improves performance.

Before:
```
test maximum_matching_bigger    ... bench:       1,044.74 ns/iter (+/- 8.21)
test maximum_matching_bipartite ... bench:         250.49 ns/iter (+/- 4.48)
test maximum_matching_full      ... bench:         129.69 ns/iter (+/- 1.10)
test maximum_matching_huge      ... bench:     847,857.70 ns/iter (+/- 25,993.22)
```

After:
```
test maximum_matching_bigger    ... bench:         984.27 ns/iter (+/- 31.05)
test maximum_matching_bipartite ... bench:         235.41 ns/iter (+/- 4.40)
test maximum_matching_full      ... bench:         130.42 ns/iter (+/- 1.04)
test maximum_matching_huge      ... bench:     798,815.20 ns/iter (+/- 12,723.48)
```